refactoring: renamed db package to rocksdb
