+++ src/main/java/com/github/lukaszbudnik/roxdb/rocksdb/Item.java
@@ -1,4 +1,4 @@
-package com.github.lukaszbudnik.roxdb.db;
+package com.github.lukaszbudnik.roxdb.rocksdb;
 
 import java.util.Map;
