+++ src/main/java/com/github/lukaszbudnik/roxdb/rocksdb/Key.java
@@ -1,4 +1,4 @@
-package com.github.lukaszbudnik.roxdb.db;
+package com.github.lukaszbudnik.roxdb.rocksdb;
 
 public record Key(String partitionKey, String sortKey) {
   public String toStorageKey() {
